+++ src/arize/experimental/integrations/whylabs_vanguard_ingestion/__init__.py
@@ -1,0 +1,8 @@
+from arize.experimental.integrations.whylabs_vanguard_ingestion.client import (
+    IntegrationClient,
+)
+from arize.experimental.integrations.whylabs_vanguard_ingestion.generator import (
+    WhylabsProfileAdapter,
+)
+
+__all__ = ["WhylabsProfileAdapter", "IntegrationClient"]